requirements.txt
tensorflow-gpu==2.5
jupyter
pillow
matplotlib
opencv-python
xlwt